# config-api/middleware_collections/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pour gérer les emplois du temps et alertes",
        "title": "Timetable Alerter API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/alerts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Liste toutes les alertes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Alert"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Crée une alerte",
                "parameters": [
                    {
                        "description": "Nouvelle alerte",
                        "name": "alert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Alert"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Alert"
                        }
                    },
                    "400": {
                        "description": "JSON invalide ou UUID incorrect",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/alerts/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Récupère une alerte",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de l'alerte",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Alert"
                        }
                    },
                    "404": {
                        "description": "Alerte non trouvée",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "alerts"
                ],
                "summary": "Met à jour une alerte",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de l'alerte",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Alerte mise à jour",
                        "name": "alert",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Alert"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Mise à jour réussie",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ID ou JSON invalide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "alerts"
                ],
                "summary": "Supprime une alerte",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de l'alerte",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Suppression réussie",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ID invalide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/resources": {
            "get": {
                "description": "Retourne tous les emplois du temps configurés",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Liste toutes les resources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Resource"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Ajoute un emploi du temps dans la base",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Crée une resource",
                "parameters": [
                    {
                        "description": "Données de la resource",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    },
                    "400": {
                        "description": "JSON invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/resources/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Récupère une resource par ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la resource",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    },
                    "404": {
                        "description": "Resource non trouvée",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "resources"
                ],
                "summary": "Met à jour une resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la resource",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Données mises à jour",
                        "name": "resource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Resource"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Mise à jour réussie",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "resources"
                ],
                "summary": "Supprime une resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la resource",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Suppression réussie",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ID invalide",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Alert": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "description": "email à notifier",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "timetable_id": {
                    "description": "clé étrangère vers Timetable",
                    "type": "string"
                },
                "when": {
                    "description": "always, room, added, removed",
                    "type": "string"
                }
            }
        },
        "models.Resource": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "description": "ex: \"M1 Groupe 1\"",
                    "type": "string"
                },
                "url": {
                    "description": "URL pour récupérer les cours",
                    "type": "string"
                }
            }
        }
    }
}