# config-api/middleware_collections/internal/controllers/collections/resource/post_resource.go
package resource

import (
	"encoding/json"
	resouceService "middleware_collections/internal/services/collections"
	"net/http"
)

// PostResourceHandler godoc
// @Summary      Crée une resource
// @Description  Ajoute un emploi du temps dans la base
// @Tags         resources
// @Accept       json
// @Produce      json
// @Param        resource  body      models.Resource  true  "Données de la resource"
// @Success      201       {object}  models.Resource
// @Failure      400       {string}  string "JSON invalide"
// @Failure      500       {string}  string "Erreur serveur"
// @Router       /resources [post]
func PostResourceHandler(w http.ResponseWriter, r *http.Request) {
	var body struct {
		Name string `json:"name"`
		URL  string `json:"url"`
	}

	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		http.Error(w, "Format JSON invalide", http.StatusBadRequest)
		return
	}

	resource, err := resouceService.CreateResource(body.Name, body.URL)
	if err != nil {
		http.Error(w, "Erreur serveur", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(resource)
}